refactor: centralize constants
