+++ src/als.h
@@ -1,0 +1,5 @@
+#pragma once
+#include "globals.h"
+
+uint32_t measureLuxWithoutBacklight();
+void updateBacklightLevel();
